requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==4.4.1
django-coverage-plugin==1.5.0
Sphinx==1.6.1
django-extensions==1.7.9
Werkzeug==0.12.2
django-test-plus==1.0.17
factory_boy==2.8.1

django-debug-toolbar==1.8

# improved REPL
ipython==5.3.0 # pyup: >=5.3.0,<6.0.0
ipdb==0.10.3

pytest-django==3.1.2
pytest-sugar==0.8.0

# Own stuff
django-rosetta==0.7.13


# REST-Framework

djangorestframework==3.6.3
markdown==2.1.0
django-filter==1.0.1
django-stdimage-serializer
